Refactor - place common fixture in conftest.py
