+++ bia-ingest-shared-models/test/test_bia_ingest_cli.py
@@ -1,6 +1,3 @@
-from unittest.mock import Mock
-from typing import Dict
-from pathlib import Path
 from typer.testing import CliRunner
 from bia_ingest_sm import cli
 from bia_ingest_sm.conversion.utils import settings
@@ -36,30 +33,20 @@ for expected_objects in expected_objects_dict.values():
         n_expected_objects += 1
 
 
-def test_cli_writes_expected_files(monkeypatch, tmp_path, test_submission):
+def test_cli_writes_expected_files(
+    monkeypatch, tmp_path, test_submission, mock_request_get
+):
     monkeypatch.setattr(settings, "bia_data_dir", str(tmp_path))
 
     def _load_submission(accession_id: str) -> biostudies.Submission:
         return test_submission
 
     monkeypatch.setattr(cli, "load_submission", _load_submission)
-    monkeypatch.setattr(biostudies, "load_submission", _load_submission)
-
-    def mock_request_get(flist_url: str) -> Dict[str, str]:
-        data_dir = Path(__file__).parent / "data"
-        path_to_load = data_dir / Path(flist_url).name
-        return_value = Mock()
-        return_value.status_code = 200
-        return_value.content = path_to_load.read_text()
-        return return_value
-
-    monkeypatch.setattr(biostudies.requests, "get", mock_request_get)
 
     result = runner.invoke(cli.app, ["ingest", accession_id])
     assert result.exit_code == 0
 
     files_written = [f for f in tmp_path.rglob("*.json")]
-    print(f"Output = {result.output.rstrip()}")
 
     assert len(files_written) == n_expected_objects
