+++ bia-ingest-shared-models/test/test_file_reference.py
@@ -5,30 +5,11 @@ now has a different pattern of creation from other artefacts i.e.
 it now needs a submitted dataset
 """
 
-from typing import Dict
-from pathlib import Path
-from unittest.mock import Mock
-import pytest
 from . import utils
 from bia_ingest_sm.conversion import (
     file_reference,
 )
-from bia_ingest_sm.biostudies import requests, File
-
-
-@pytest.fixture
-def mock_request_get(monkeypatch):
-    """Requests.get mocked to read file from disk"""
-
-    def _mock_request_get(flist_url: str) -> Dict[str, str]:
-        data_dir = Path(__file__).parent / "data"
-        path_to_load = data_dir / Path(flist_url).name
-        return_value = Mock()
-        return_value.status_code = 200
-        return_value.content = path_to_load.read_text()
-        return return_value
-
-    monkeypatch.setattr(requests, "get", _mock_request_get)
+from bia_ingest_sm.biostudies import File
 
 
 # Get second study component as dataset in submission
@@ -49,17 +30,19 @@ def test_get_file_reference_for_submission_dataset(test_submission, result_summa
         accession_id=test_submission.accno,
         submission_dataset=datasets_in_submission[0],
         files_in_file_list=files_in_filelist,
-        result_summary=result_summary
+        result_summary=result_summary,
     )
     assert created == expected
 
 
-def test_create_file_reference_for_study_component(test_submission, caplog, result_summary, mock_request_get):
-
+def test_create_file_reference_for_study_component(
+    test_submission, caplog, result_summary, mock_request_get
+):
     expected = {datasets_in_submission[0].title_id: utils.get_test_file_reference()}
     created = file_reference.get_file_reference_by_dataset(
-        test_submission, datasets_in_submission=datasets_in_submission, result_summary=result_summary
-
+        test_submission,
+        datasets_in_submission=datasets_in_submission,
+        result_summary=result_summary,
     )
     assert created == expected
 
@@ -77,7 +60,11 @@ def test_create_file_reference_for_study_component_when_no_matching_sc_in_file_l
     dataset = utils.get_test_experimental_imaging_dataset()[0]
     dataset.title_id = "Test name not in file list"
     created = file_reference.get_file_reference_by_dataset(
-        test_submission, datasets_in_submission=[dataset,], result_summary=result_summary
+        test_submission,
+        datasets_in_submission=[
+            dataset,
+        ],
+        result_summary=result_summary,
     )
 
     assert created is None
